Migrated stringArray to its own module.
