+++ rust/arrow/src/array/raw_pointer.rs
@@ -1,3 +1,5 @@
+use crate::memory;
+
 pub(super) struct RawPtrBox<T> {
     inner: *const T,
 }
@@ -14,3 +16,11 @@ impl<T> RawPtrBox<T> {
 
 unsafe impl<T> Send for RawPtrBox<T> {}
 unsafe impl<T> Sync for RawPtrBox<T> {}
+
+pub(super) fn as_aligned_pointer<T>(p: *const u8) -> *const T {
+    assert!(
+        memory::is_aligned(p, std::mem::align_of::<T>()),
+        "memory is not aligned"
+    );
+    p as *const T
+}
